feat: add new info structure desktop header
- with feature toggle
- add storybook & new component stories
