+++ packages/universal-header/.storybook/main.js
@@ -1,0 +1,27 @@
+const path = require('path')
+
+module.exports = {
+  stories: [
+    '../src/**/**/*.stories.mdx',
+    '../src/**/**/*.stories.@(js|jsx|ts|tsx)',
+  ],
+  addons: [
+    '@storybook/addon-links',
+    '@storybook/addon-essentials',
+    '@storybook/addon-interactions',
+    '@storybook/addon-viewport',
+    '@storybook/addon-backgrounds',
+  ],
+  framework: '@storybook/react',
+  webpackFinal: async (config, { configType }) => {
+    config.resolve.alias['@twreporter/core/lib'] = path.resolve(
+      __dirname,
+      '../../core/src/'
+    )
+    config.resolve.alias['@twreporter/react-components/lib'] = path.resolve(
+      __dirname,
+      '../../react-components/src/'
+    )
+    return config
+  },
+}
