+++ packages/universal-header/src/components/icons.js
@@ -1,240 +1,165 @@
-import HeaderContext from '../contexts/header-context'
-import Link from './customized-link'
-import PropTypes from 'prop-types'
-import React from 'react'
-import SearchBox from './search-box'
-import fonts from '../constants/fonts'
-import linkUtils from '../utils/links'
+import React, { useState, useContext } from 'react'
 import querystring from 'querystring'
 import styled from 'styled-components'
-import themeUtils from '../utils/theme'
+// context
+import HeaderContext from '../contexts/header-context'
+// utils
+import linkUtils from '../utils/links'
 // @twreporter
-import mq from '@twreporter/core/lib/utils/media-query'
-// lodash
-import map from 'lodash/map'
-
-const _ = {
-  map,
-}
-
-const styles = {
-  iconContainerSize: 3, // em
-}
+import Link from '@twreporter/react-components/lib/customized-link'
+import { IconButton } from '@twreporter/react-components/lib/button'
+import { Member, Search, Bookmark } from '@twreporter/react-components/lib/icon'
+import { Dialog } from '@twreporter/react-components/lib/card'
+import { SearchBar } from '@twreporter/react-components/lib/input'
+import { useOutsideClick } from '@twreporter/react-components/lib/hook'
 
 const IconsContainer = styled.div`
-  position: relative;
-  display: table;
-  ${mq.mobileOnly`
-    display: none;
-  `}
+  display: flex;
 `
 
 const IconContainer = styled.div`
-  font-size: ${fonts.size.base};
-  cursor: pointer;
-  display: table-cell;
-  width: ${styles.iconContainerSize}em;
-  height: ${styles.iconContainerSize}em;
-  line-height: 1;
   position: relative;
-  opacity: ${props => (props.isSearchOpened ? '0' : '1')};
-  transition: opacity 600ms ease;
+  margin-right: 16px;
   a {
     display: flex;
   }
-  svg {
-    opacity: 1;
-    position: absolute;
-    height: 100%;
-    top: 0;
-    left: 30%;
-    z-index: 1;
-  }
 `
 
-const ShowOnHover = styled.div`
-   display: none;
-
-   ${IconContainer}:hover & {
-     display: block;
-   }
+const Container = styled.div`
+  opacity: ${props => (props.isSearchOpened ? '0' : '1')};
+  transition: opacity 300ms ease;
 `
 
-const HideOnHover = styled.div`
-   display: block;
-
-   ${IconContainer}:hover & {
-     display: none;
-   }
+const LogContainer = styled.div`
+  display: flex;
+  flex-direction: column;
 `
 
-class Icons extends React.PureComponent {
-  static propTypes = {
-    services: PropTypes.array,
-  }
+const SearchContainer = styled.div`
+  opacity: ${props => (props.isSearchOpened ? '1' : '0')};
+  transition: opacity 300ms ease;
+  position: absolute;
+  right: 0;
+  top: -16px;
+  z-index: ${props => (props.isSearchOpened ? 999 : -1)};
+`
 
-  static defaultProps = {
-    services: [],
-  }
+const StyledDialog = styled(Dialog)`
+  opacity: ${props => (props.showDialog ? '1' : '0')};
+  transition: opacity 100ms;
+  position: absolute;
+  top: 24px;
+  right: -16px;
+  width: max-content;
+  cursor: pointer;
+`
 
-  constructor(props) {
-    super(props)
-    this.state = {
-      isSearchOpened: false,
-    }
-    this._closeSearchBox = this._closeSearchBox.bind(this)
-    this._handleClickSearch = this._handleClickSearch.bind(this)
-  }
-
-  _closeSearchBox() {
-    this.setState({
-      isSearchOpened: false,
-    })
-  }
-
-  _handleClickSearch(e) {
+const LogInOutIcon = () => {
+  const [showDialog, setShowDialog] = useState(false)
+  const { releaseBranch, theme, isAuthed } = useContext(HeaderContext)
+  const onClickIcon = e => {
     e.preventDefault()
-    this.setState({
-      isSearchOpened: true,
-    })
-  }
 
-  _handleLogIconClick = (e, isAuthed, releaseBranch) => {
+    if (isAuthed) {
+      setShowDialog(!showDialog)
+      return
+    }
+    const redirectURL = window.location.href
+    const query = querystring.stringify({ destination: redirectURL })
+    window.location = linkUtils.getLogoutLink(releaseBranch).to + '?' + query
+  }
+  const onClickLogOut = e => {
     e.preventDefault()
 
     const redirectURL = window.location.href
-    const query = querystring.stringify({
-      destination: redirectURL,
-    })
-    if (isAuthed) {
-      window.location = linkUtils.getLogoutLink(releaseBranch).to + '?' + query
-      return
-    }
+    const query = querystring.stringify({ destination: redirectURL })
     window.location = linkUtils.getLoginLink(releaseBranch).to + '?' + query
   }
+  const closeDialog = () => setShowDialog(false)
+  const ref = useOutsideClick(closeDialog)
 
-  _prepareIconJsx(service) {
-    const { isSearchOpened } = this.state
-    const Login = (
-      <IconContainer isSearchOpened={isSearchOpened} key="login">
-        <HeaderContext.Consumer>
-          {({ releaseBranch, theme }) => {
-            const [LogInIcon, LogInHoverIcon] = themeUtils.selectServiceIcons(theme).login
-            return (
-              <Link
-                onClick={e => {
-                  this._handleLogIconClick(e, false, releaseBranch)
-                }}
-              >
-                <HideOnHover>
-                  <LogInIcon />
-                </HideOnHover>
-                <ShowOnHover>
-                  <LogInHoverIcon />
-                </ShowOnHover>
-              </Link>
-            )
-          }}
-        </HeaderContext.Consumer>
-      </IconContainer>
-    )
-    const Logout = (
-      <IconContainer isSearchOpened={isSearchOpened} key="logout">
-        <HeaderContext.Consumer>
-          {({ releaseBranch, theme }) => {
-            const [LogOutIcon, LogOutHoverIcon] = themeUtils.selectServiceIcons(theme).logout
-            return (
-              <Link
-                onClick={e => {
-                  this._handleLogIconClick(e, true, releaseBranch)
-                }}
-              >
-                <HideOnHover>
-                  <LogOutIcon />
-                </HideOnHover>
-                <ShowOnHover>
-                  <LogOutHoverIcon />
-                </ShowOnHover>
-              </Link>
-            )
-          }}
-        </HeaderContext.Consumer>
-      </IconContainer>
-    )
-    const Search = (
-      <React.Fragment key="search">
-        <IconContainer
-          onClick={this._handleClickSearch}
-          isSearchOpened={isSearchOpened}
-        >
-          <HeaderContext.Consumer>
-            {({ theme }) => {
-              const [SearchIcon, SearchHoverIcon] = themeUtils.selectServiceIcons(theme).search
-              return (
-                <React.Fragment>
-                  <HideOnHover>
-                    <SearchIcon />
-                  </HideOnHover>
-                  <ShowOnHover>
-                    <SearchHoverIcon />
-                  </ShowOnHover>
-                </React.Fragment>
-              )
-            }}
-          </HeaderContext.Consumer>
-        </IconContainer>
-        <SearchBox
-          isSearchOpened={isSearchOpened}
-          closeSearchBox={this._closeSearchBox}
+  const Icon = <Member releaseBranch={releaseBranch} />
+  return (
+    <IconContainer key="login">
+      <LogContainer onClick={onClickIcon} ref={ref}>
+        <IconButton iconComponent={Icon} theme={theme} />
+        <StyledDialog
+          text="登出"
+          size="L"
+          showDialog={showDialog}
+          onClick={onClickLogOut}
         />
-      </React.Fragment>
-    )
-    const Bookmark = (
-      <IconContainer isSearchOpened={isSearchOpened} key="bookmark">
-        <HeaderContext.Consumer>
-          {({ releaseBranch, isLinkExternal, theme }) => {
-            const [BookmarkIcon, BookmarkHoverIcon] = themeUtils.selectServiceIcons(theme).bookmark
-            const link = linkUtils.getBookmarksLink(
-              isLinkExternal,
-              releaseBranch
-            )
-            return (
-              <Link {...link}>
-                <HideOnHover>
-                  <BookmarkIcon />
-                </HideOnHover>
-                <ShowOnHover>
-                  <BookmarkHoverIcon />
-                </ShowOnHover>
-              </Link>
-            )
-          }}
-        </HeaderContext.Consumer>
-      </IconContainer>
+      </LogContainer>
+    </IconContainer>
+  )
+}
 
-    )
-    switch(service) {
-      case 'login':
-        return Login
-      case 'logout':
-        return Logout
-      case 'search':
-        return Search
-      case 'bookmarks':
-        return Bookmark
-      default:
-        return null;
+const SearchIcon = () => {
+  const [isSearchOpened, setSearchOpened] = useState(false)
+  const { isLinkExternal, releaseBranch, theme } = useContext(HeaderContext)
+
+  const closeSearchBox = () => {
+    setSearchOpened(false)
+  }
+  const handleClickSearch = e => {
+    e.preventDefault()
+    setSearchOpened(true)
+    if (!ref.current) {
+      return
+    }
+    const input = ref.current.getElementsByTagName('INPUT')[0]
+    if (input) {
+      input.focus()
     }
   }
-
-  render() {
-    const { services } = this.props;
-    return (
-      <IconsContainer>
-        { _.map(services, service => this._prepareIconJsx(service.key)) }
-      </IconsContainer>
-    )
+  const onSearch = keywords => {
+    if (!window) {
+      return
+    }
+    window.location = `${
+      linkUtils.getSearchLink(isLinkExternal, releaseBranch).to
+    }?q=${keywords}`
   }
+
+  const Icon = <Search releaseBranch={releaseBranch} />
+  const ref = useOutsideClick(closeSearchBox)
+  return (
+    <IconContainer ref={ref} key="search">
+      <Container onClick={handleClickSearch} isSearchOpened={isSearchOpened}>
+        <IconButton iconComponent={Icon} theme={theme} />
+      </Container>
+      <SearchContainer isSearchOpened={isSearchOpened}>
+        <SearchBar
+          placeholder="關鍵字搜尋"
+          theme={theme}
+          onClose={closeSearchBox}
+          onSearch={onSearch}
+        />
+      </SearchContainer>
+    </IconContainer>
+  )
 }
+
+const BookmarkIcon = () => {
+  const { releaseBranch, isLinkExternal, theme } = useContext(HeaderContext)
+  const link = linkUtils.getBookmarksLink(isLinkExternal, releaseBranch)
+  const Icon = <Bookmark releaseBranch={releaseBranch} />
+
+  return (
+    <IconContainer key="bookmark">
+      <Link {...link}>
+        <IconButton iconComponent={Icon} theme={theme} />
+      </Link>
+    </IconContainer>
+  )
+}
+
+const Icons = () => (
+  <IconsContainer>
+    <SearchIcon />
+    <BookmarkIcon />
+    <LogInOutIcon />
+  </IconsContainer>
+)
 
 export default Icons
