+++ packages/universal-header/src/index.js
@@ -1,5 +1,9 @@
 import StandaloneHeader from './standalone-header'
-import Header from './containers/header'
+import HeaderNew from './containers/header'
+import HeaderOld from './containers/header-old'
+// feature toggle
+import { ENABLE_NEW_INFO_ARCH } from '@twreporter/core/lib/constants/feature-flag'
+export const Header = ENABLE_NEW_INFO_ARCH ? HeaderNew : HeaderOld
 
 export default {
   Header,
